feat: Add application initialization and running logic #main
Add new files to initialize and run the application. Three main files
are added which include `app.go`, `service_provider.go`, and `main.go`.
In `app.go`, the structure of the application is defined and methods
are added for initialization and running of the app. In
`service_provider.go`, service providers are initialized. And finally,
in `main.go`, the app is created and run.